разбили весь код на модули + завели цикл в main вместо многократного вызова display_menu
